requirements.txt
APScheduler==3.11.2
bcrypt==5.0.0
blinker==1.9.0
click==8.3.2
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.3.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
PyMySQL==1.1.2
qrcode[pil]>=8.0
python-dotenv==1.2.2
SQLAlchemy==2.0.49
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
Werkzeug==3.1.8
WTForms==3.2.1
cryptography>=42.0.0
python-barcode[images]>=0.15.1
